pyproject.toml
[tool.poetry]
name = "musharing-chatbot"
version = "0.1.0"
description = ""
authors = ["CDFMLR <cdfmlr@163.com>"]
readme = "README.md"
packages = [{include = "musharing_chatbot"}]

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
# chatterbot = {url = "https://github.com/RaSan147/ChatterBot_update/archive/refs/heads/master.zip"}
chatterbot = {url = "https://github.com/RaSan147/ChatterBot_update/archive/265bc0511563f2f4b927141d667d9361f579454f.zip"}
grpcio = "^1.51.1"
protobuf = "^4.22.0"
grpcio-reflection = "^1.51.1"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
secondary = false


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
